test(angular-query/inject-query): move type assertions from the runtime test to 'test-d'
